refactor component props into model folder
