+++ packages/tinacms/src/toolkit/fields/components/reference/model/reference-link-props.ts
@@ -1,0 +1,19 @@
+import { TinaCMS } from '@toolkit/tina-cms'
+
+export interface ReferenceLinkProps {
+  cms: TinaCMS
+  input: any
+}
+
+export type Document = {
+  _sys: {
+    collection: {
+      name: string
+    }
+    breadcrumbs: string[]
+  }
+}
+
+export interface Response {
+  node: Document
+}
